requirements.txt
fastapi[all]
sqlalchemy[asyncio]
databases[mysql]
pydantic_settings
alembic
gunicorn
uvicorn[standard]
asyncmy
aioredis
passlib
python-jose[cryptography]
bcrypt
motor
pydantic
aiokafka
websockets
aioboto3==13.3.0
accelerate
pdf2image
pymilvus
colpali_engine
tqdm
pillow
qwen-vl-utils[decord]==0.0.8
torchvision
openai
tenacity
#git+https://github.com/illuin-tech/colpali